+++ changelog.md
@@ -1,4 +1,7 @@
 # eT v1.6.0
+### Features
+- Print tool for eigen and linear Davidson solvers. eT-program/eT!983
+
 ### Optimization
 - Coupled cluster Fock matrix construction is N^4 scaling. eT-program/eT!973
 - Reduce memory usage for MLCC energy calculation (reusing CCS routine). eT-program/eT!974